Update BulkUpsert with batch size implementation

+++ SaveChangesResolverTest/ResolverTest.cs
@@ -6,6 +6,7 @@ using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.IdentityModel.Tokens;
+using System.Diagnostics;
 using System.Reflection;
 
 namespace SaveChangesResolverTest
@@ -14,25 +15,32 @@ namespace SaveChangesResolverTest
     {
         private EntityProjectContext context;
         private SaveChangesResolver resolver;
+        private bool IntegrationTesting = true;
 
         [OneTimeSetUp]
         public void OneTimeSetUp()
         {
-            DbContextOptionsBuilder<EntityProjectContext> optionsBuilder = new DbContextOptionsBuilder<EntityProjectContext>();
-            var connection = new SqliteConnection("Data Source=InMemorySample;Mode=Memory;Cache=Shared");
-            connection.Open();
-            optionsBuilder.UseSqlite(connection);
-
-            context = new EntityProjectContext(optionsBuilder.Options);
-            context.Database.EnsureDeleted();
-            context.Database.EnsureCreated();
-            resolver = new SaveChangesResolver(context);
+            if (!IntegrationTesting)
+            {
+                DbContextOptionsBuilder<EntityProjectContext> optionsBuilder = new DbContextOptionsBuilder<EntityProjectContext>();
+                var connection = new SqliteConnection("Data Source=InMemorySample;Mode=Memory;Cache=Shared");
+                connection.Open();
+                optionsBuilder.UseSqlite(connection);
+                context = new EntityProjectContext(optionsBuilder.Options);
+                context.Database.EnsureDeleted();
+                context.Database.EnsureCreated();
+            }
+            else
+            {
+                context = new EntityProjectContext();
+            }
         }
 
         [SetUp]
         public void SetUp()
         {
             hashSetCount = 0;
+            resolver = new SaveChangesResolver(context);
         }
 
         [Test]
@@ -51,14 +59,6 @@ namespace SaveChangesResolverTest
         public void PrintDependencyOrder()
         {
             Console.Write(resolver.PrintDependencyOrder());
-        }
-        
-        [Test]
-        public void DeferUpsert()
-        {
-            resolver.DeferUpsert(new Table1() { Col1_PK = 1, Col2 = "1" });
-            resolver.DeferUpsert(new Table1() { Col1_PK = 1, Col2 = "2" });
-            Console.Write(resolver.PrintTotalItemsDeferred());
         }
 
         private void AddToHashSet(HashSet<object> hashset, object obj)
@@ -125,13 +125,59 @@ namespace SaveChangesResolverTest
             AddToHashSet(hashset, b);
             Assert.AreEqual(5, hashset.Count());
         }
-        
+
         [Test]
-        public void BulkUpsert()
+        public void DeferUpsert_Print()
         {
-            DeferUpsert();
+            resolver.DeferUpsert(new Table1() { Col1_PK = 1, Col2 = "1" });
+            resolver.DeferUpsert(new Table1() { Col1_PK = 1, Col2 = "2" });
+            resolver.DeferUpsert(new Table1() { Col1_PK = 2, Col2 = "2" });
+            resolver.DeferUpsert(new Table2() { Col1_PK = 3, Col2_FK = 4 });
+            
+            Console.WriteLine(resolver.PrintTotalItemsDeferred(false));
+            
+            Assert.AreEqual(2, resolver.GetTotalItemsDeferredByType(typeof(Table1)));
+            Assert.AreEqual(1, resolver.GetTotalItemsDeferredByType(typeof(Table2)));
+            
+            Console.WriteLine(resolver.PrintTotalItemsDeferred(true));
+        }
+
+        [Test]
+        public void BulkDeferUpsert_Print()
+        {
+            BulkDeferUpsert(19000);
+            Console.WriteLine(resolver.PrintTotalItemsDeferred(true));
+        }
+
+        [TestCase(100000)]
+        public void BulkDeferUpsert(int totalInserts)
+        {
+            int j = 0;
+            for (int i = 0; i < totalInserts; i++)
+            {
+                if (j == 25)
+                    j = 0;
+                resolver.DeferUpsert(new Table1() { Col1_PK = i, Col2 = i.ToString(), Col4 = "" });
+                resolver.DeferUpsert(new Table2() { Col1_PK = j, Col2_FK = j, Col3_Value = "", Col4_Extra = "", Col5_Extra = i-j });
+                j++;
+            }
+            Console.WriteLine(resolver.PrintTotalItemsDeferred(false));
+
+            Console.WriteLine("Table1 items deferred: " + resolver.GetTotalItemsDeferredByType(typeof(Table1)));
+            Console.WriteLine("Table2 items deferred: " + resolver.GetTotalItemsDeferredByType(typeof(Table2)));
+            Console.WriteLine();
+        }
+
+        [TestCase(1000)]
+        public void BulkUpsert(int bulksize)
+        {
+            Console.WriteLine("BEFORE");
+            BulkDeferUpsert(bulksize);
+            
             resolver.BulkUpsert();
 
+            Console.WriteLine("AFTER");
+            BulkDeferUpsert(0);
         }
     }
 }
